+++ src/index.ts
@@ -3,10 +3,11 @@ import { pipe } from 'fp-ts/lib/function'
 import * as E from 'fp-ts/lib/Either'
 import {
   IBeeConfig, IEntityContentJson,
-  IBeeLoader, IBeeOptions, ILoadConfig, ILoadStageMode, IUrlConfig, LoadWorkspaceOptions,
-  IExecParams, IUpdateToken
+  IBeeLoader, IBeeOptions, ILoadConfig,
+  ILoadStageMode, IUrlConfig, LoadWorkspaceOptions,
+  IToken, BeeSaveOptions, IExecParams
 } from './types/bee'
-import beeActions from './utils/Constants'
+import beeActions, { mockedEmptyToken } from './utils/Constants'
 import { fetchToken } from './services/api'
 import { eitherCanExecuteAction, eitherCheckJoinParams, eitherCheckStartParams } from './utils/utils'
 import * as beeTypes from './types/bee'
@@ -53,16 +54,16 @@ const {
 } = beeActions
 
 class Bee {
-  token: string
+  token: IToken
   bee: any // todo delete
   instance: any
 
   constructor(
-    token?: string, urlConfig: IUrlConfig = { authUrl: API_AUTH_URL, beePluginUrl: BEEJS_URL }
+    token?: IToken, urlConfig: IUrlConfig = { authUrl: API_AUTH_URL, beePluginUrl: BEEJS_URL }
   ) {
     beeLoaderUrl = urlConfig
     this.bee = (call) => load(() => call())
-    this.token = token || ''
+    this.token = token || mockedEmptyToken
     this.instance = null   
   }
 
@@ -72,7 +73,7 @@ class Bee {
     urlConfig:IUrlConfig = { authUrl: API_AUTH_URL, beePluginUrl: BEEJS_URL }
   ) => {
     beeLoaderUrl = urlConfig;
-    if (this.token) {
+    if (this.token && this.token.access_token) {
       throw new Error('Toker already declared')
     }
 
@@ -85,13 +86,13 @@ class Bee {
 
   start = (
     config: IBeeConfig,
-    template: IEntityContentJson,
+    template: IEntityContentJson | object, //user can pass an empty object in some specific cases
     bucketDir?: string, 
     options?: IBeeOptions
   ) => {
     const { bee, token } = this    
     return pipe(
-      eitherCheckStartParams(config, template, token),
+      eitherCheckStartParams(config, token),
       E.fold(
         ({ message }) => new Promise(reject => reject(message)),
         () => new Promise(resolve => {
@@ -141,11 +142,11 @@ class Bee {
     )
   }
 
-  load = (template: Record<string, unknown>) => this.executeAction(LOAD, template)
+  load = (template: IEntityContentJson) => this.executeAction(LOAD, template)
 
   loadRows = () => this.executeAction(LOAD_ROWS)
 
-  save = () => this.executeAction(SAVE)
+  save = (options?: BeeSaveOptions) => this.executeAction(SAVE, options)
 
   saveAsTemplate = () => this.executeAction(SAVE_AS_TEMPLATE)
 
@@ -163,7 +164,7 @@ class Bee {
 
   showComment = (comment) => this.executeAction(SHOW_COMMENT, comment) 
 
-  reload = (template: Record<string, unknown>, options?: IBeeOptions) => this.executeAction(RELOAD, template, options)
+  reload = (template: IEntityContentJson, options?: IBeeOptions) => this.executeAction(RELOAD, template, options)
 
   loadWorkspace = (type: LoadWorkspaceOptions) => this.executeAction(LOAD_WORKSPACE, type)
 
@@ -174,7 +175,7 @@ class Bee {
   /** EXPERIMENTAL DO NOT use it in production */
   exec = (args: IExecParams) => this.executeAction(EXEC, { cmd: args.cmd, ...args.params})
 
-  updateToken = (updateTokenArgs: IUpdateToken) => this.executeAction(UPDATE_TOKEN, updateTokenArgs)
+  updateToken = (updateTokenArgs: IToken) => this.executeAction(UPDATE_TOKEN, updateTokenArgs)
 
 }